Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Yes, just for awareness. Does it make sense ? As it's a tempest
 > plugin..

I saw it in multiple places while reviewing tempest-plugin reqchecks, and after seeing this differnce there for 2 years and chandan said it's fine to keep it as it is. As tempest maintainers are aware of it we can skip it i think.

Patch-set: 1
Reviewer: Gerrit User 403 <403@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0